Addressed more of the feedback
