+++ community-care-eligibility/pom.xml
@@ -11,7 +11,7 @@
   <version>1.0.3-SNAPSHOT</version>
   <packaging>jar</packaging>
   <properties>
-    <ee-artifacts.version>1.1.36</ee-artifacts.version>
+    <ee-artifacts.version>1.1.35</ee-artifacts.version>
     <jacoco.coverage>0.50</jacoco.coverage>
   </properties>
   <dependencies>
