+++ community-care-eligibility/src/main/java/gov/va/api/health/communitycareeligibility/service/VaFacilitiesResponse.java
@@ -1,15 +1,22 @@
 package gov.va.api.health.communitycareeligibility.service;
 
+import com.fasterxml.jackson.annotation.JsonAutoDetect;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonProperty;
 import java.util.ArrayList;
 import java.util.List;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Data
 @Builder
+@JsonIgnoreProperties(ignoreUnknown = true)
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
+@JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
 public final class VaFacilitiesResponse {
   private List<VaFacility> data;
 
@@ -23,10 +30,111 @@ public final class VaFacilitiesResponse {
 
   @Data
   @Builder
+  @JsonIgnoreProperties(ignoreUnknown = true)
+  @NoArgsConstructor(access = AccessLevel.PRIVATE)
   @AllArgsConstructor(access = AccessLevel.PRIVATE)
+  @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
   public static final class VaFacility {
     private String id;
 
     private VaFacilityAttributes attributes;
+  }
+
+  @Data
+  @Builder
+  @JsonIgnoreProperties(ignoreUnknown = true)
+  @NoArgsConstructor(access = AccessLevel.PRIVATE)
+  @AllArgsConstructor(access = AccessLevel.PRIVATE)
+  @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
+  public static final class VaFacilityAttributes {
+    private String name;
+
+    private Double lat;
+
+    @JsonProperty("long")
+    private Double longg;
+
+    private Address address;
+
+    private Phone phone;
+
+    @JsonProperty("wait_times")
+    private WaitTimes waitTimes;
+
+    @Data
+    @Builder
+    @JsonIgnoreProperties(ignoreUnknown = true)
+    @NoArgsConstructor(access = AccessLevel.PRIVATE)
+    @AllArgsConstructor(access = AccessLevel.PRIVATE)
+    @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
+    public static final class Address {
+      private PhysicalAddress physical;
+    }
+
+    @Data
+    @Builder
+    @JsonIgnoreProperties(ignoreUnknown = true)
+    @NoArgsConstructor(access = AccessLevel.PRIVATE)
+    @AllArgsConstructor(access = AccessLevel.PRIVATE)
+    @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
+    public static final class Phone {
+      private String main;
+    }
+
+    @Data
+    @Builder
+    @JsonIgnoreProperties(ignoreUnknown = true)
+    @NoArgsConstructor(access = AccessLevel.PRIVATE)
+    @AllArgsConstructor(access = AccessLevel.PRIVATE)
+    @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
+    public static final class PhysicalAddress {
+      private String zip;
+
+      private String city;
+
+      private String state;
+
+      @JsonProperty("address_1")
+      private String address1;
+
+      @JsonProperty("address_2")
+      private String address2;
+
+      @JsonProperty("address_3")
+      private String address3;
+    }
+
+    @Data
+    @Builder
+    @JsonIgnoreProperties(ignoreUnknown = true)
+    @NoArgsConstructor(access = AccessLevel.PRIVATE)
+    @AllArgsConstructor(access = AccessLevel.PRIVATE)
+    @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
+    public static final class WaitTime {
+      private String service;
+
+      @JsonProperty("new")
+      private Integer neww;
+
+      private Integer established;
+    }
+
+    @Data
+    @Builder
+    @JsonIgnoreProperties(ignoreUnknown = true)
+    @NoArgsConstructor(access = AccessLevel.PRIVATE)
+    @AllArgsConstructor(access = AccessLevel.PRIVATE)
+    @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
+    public static final class WaitTimes {
+      private List<WaitTime> health;
+
+      /** Health wait time. */
+      public List<WaitTime> health() {
+        if (health == null) {
+          health = new ArrayList<>();
+        }
+        return health;
+      }
+    }
   }
 }
