+++ community-care-eligibility/src/main/resources/application.properties
@@ -34,6 +34,7 @@ bing-maps.api-key=unset
 va-facilities.api-key=unset
 community-care.max-drive-time=60
 community-care.max-wait=20
+facilities.endpoint.url=unset
 ee.endpoint.url=unset
 ee.header.password=unset
 ee.header.username=unset
