+++ community-care-eligibility/src/test/java/gov/va/api/health/communitycareeligibility/service/LombokTest.java
@@ -25,6 +25,7 @@ public final class LombokTest {
   public void exerciseLombok() {
     for (Class<?> pojoClass :
         Arrays.asList(
+            VaFacilitiesResponse.VaFacility.class,
             BingResponse.class,
             BingResponse.Point.class,
             BingResponse.Resource.class,
